[FIX] S3Service 의존성 주입을 위한 코드 수정
